unsafe preprocessor module
