+++ compiler/rustc_unsafe_preprocessor/Cargo.toml
@@ -1,30 +1,15 @@
 [package]
-name = "rustc_driver_impl"
-version = "0.0.0"
+name = "rustc_unsafe_preprocessor"
+version = "0.1.0"
 edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 # tidy-alphabetical-start
 rustc_ast = { path = "../rustc_ast" }
 rustc_ast_lowering = { path = "../rustc_ast_lowering" }
-rustc_ast_passes = { path = "../rustc_ast_passes" }
-rustc_ast_pretty = { path = "../rustc_ast_pretty" }
-rustc_attr = { path = "../rustc_attr" }
-rustc_borrowck = { path = "../rustc_borrowck" }
-rustc_builtin_macros = { path = "../rustc_builtin_macros" }
-rustc_codegen_ssa = { path = "../rustc_codegen_ssa" }
-rustc_const_eval = { path = "../rustc_const_eval" }
 rustc_data_structures = { path = "../rustc_data_structures" }
-rustc_errors = { path = "../rustc_errors" }
-rustc_expand = { path = "../rustc_expand" }
-rustc_feature = { path = "../rustc_feature" }
-rustc_fluent_macro = { path = "../rustc_fluent_macro" }
-rustc_hir = { path = "../rustc_hir" }
-rustc_hir_analysis = { path = "../rustc_hir_analysis" }
-rustc_hir_pretty = { path = "../rustc_hir_pretty" }
-rustc_hir_typeck = { path = "../rustc_hir_typeck" }
-rustc_incremental = { path = "../rustc_incremental" }
-rustc_infer = { path = "../rustc_infer" }
 rustc_interface = { path = "../rustc_interface" }
 rustc_lint = { path = "../rustc_lint" }
 rustc_log = { path = "../rustc_log" }
@@ -48,7 +33,6 @@ rustc_symbol_mangling = { path = "../rustc_symbol_mangling" }
 rustc_target = { path = "../rustc_target" }
 rustc_trait_selection = { path = "../rustc_trait_selection" }
 rustc_ty_utils = { path = "../rustc_ty_utils" }
-rustc_unsafe_preprocessor = { path = "../rustc_unsafe_preprocessor" }
 serde_json = "1.0.59"
 shlex = "1.0"
 time = { version = "0.3", default-features = false, features = ["alloc", "formatting", "parsing", "macros"] }
